LayoutTests/platform/glib/imported/w3c/web-platform-tests/mathml/relations/html5-tree/display-1-expected.txt


PASS Test display math lowercase
PASS Test inline math lowercase
FAIL Test width on display=block math lowercase assert_approx_equals: math must be centered. expected 308 +/- 1 but got 58
FAIL Test display math uppercase assert_approx_equals: math must be centered. expected 308 +/- 1 but got 87
PASS Test inline math uppercase
FAIL Test width on display=block math uppercase assert_approx_equals: math must be centered. expected 308 +/- 1 but got 87
FAIL Test display math alternate_case assert_approx_equals: math must be centered. expected 308 +/- 1 but got 87
PASS Test inline math alternate_case
FAIL Test width on display=block math alternate_case assert_approx_equals: math must be centered. expected 308 +/- 1 but got 87
PASS Test dynamically setting display=block
PASS Test dynamically setting display=inline
